+++ catalogue/webapp/components/ItemRequestButton/ItemRequestButton.js
@@ -2,13 +2,13 @@
 import { useEffect, useState } from 'react';
 import { Tag } from '@weco/common/views/components/Tags/Tags';
 import { classNames, font } from '@weco/common/utils/classnames';
-import { type StacksItem } from '../WorkItemsStatus/WorkItemsStatus';
+import { type StacksItem } from '@weco/catalogue/components/WorkItemsStatus/WorkItemsStatus';
 
 import { requestItem, getUserHolds } from '../../services/stacks/requests';
 import useAuth from '@weco/common/hooks/useAuth';
 
-type ItemRequestButtonProps = {| item: StacksItem, workId: string |};
-const ItemRequestButton = ({ item, workId }: ItemRequestButtonProps) => {
+type Props = {| item: StacksItem, workId: string |};
+const ItemRequestButton = ({ item, workId }: Props) => {
   const [requestedState, setRequestedState] = useState<string>('unknown');
 
   function setRedirectCookie(workId: string, itemId: string) {
